Move getContext… function to builder package
